requirements.txt
altair==3.2.0
argh==0.26.2
astor==0.8.0
attrs==19.3.0
base58==1.0.3
blinker==1.4
boto3==1.10.5
botocore==1.13.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.1
idna==2.6
importlib-metadata==0.23
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.1.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
numpy==1.16.5
pandas==0.25.2
pathtools==0.1.2
Pillow==6.2.1
protobuf==3.10.0
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.18.4
Robinhood==1.0.1
s3transfer==0.2.1
six==1.11.0
streamlit==0.49.0
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tzlocal==2.0.0
urllib3==1.22
validators==0.14.0
watchdog==0.9.0
wincertstore==0.2
zipp==0.6.0
